MirrorSuccess
So what goes on now is that the server and client projects have a mirrored scene tree structure to each other, insofar nodes with networked RPC calls are concerned. The server proj now has a database singleton to clean that up and keep the scene tree as closely mirrored as possible without extras, and the the client proj should be much more navigable now.