lots of stuff.
Ok so, silly me right I had thought I messed up the client to server authentication after tokens distributed, since if a user has a custom client they need not even have the method to _check_token that the server calls for. That is where the timer comes into play though on the server, only give the client a set amount of time (like 4 seconds or less) to actually send their valid token or GTFO! This involved making a queue of players to get kicked and then remembering to delete them from the queue if they DO get validated as they're supposed to, etc, lots of little things.